fix: address code review findings
- Fix Galois element type mismatch: GenerateLinearTransformRotationKey and
  GenerateAndSerializeRotationKey now use C.ulong/ctypes.c_ulong instead of
  C.int/ctypes.c_int, preventing potential truncation of large Galois elements
- Remove dead NewEvaluator backward-compat classes from poly_evaluator.py and
  lt_evaluator.py (referenced deleted Scheme class, ~280 lines of unreachable code)
- Add guards on CipherTensor.decrypt(), .min(), .max(), __str__ for server-side
  tensors where encryptor is None, giving clear errors instead of AttributeError
- Fix compiler progress print formatting: {1}-{5} -> [1/5]-[5/5]

+++ orion/backend/lattigo/bindings.py
@@ -744,13 +744,13 @@ class LattigoLibrary:
 
         self.GenerateLinearTransformRotationKey = LattigoFunction(
             self.lib.GenerateLinearTransformRotationKey,
-            argtypes=[ctypes.c_int],
+            argtypes=[ctypes.c_ulong],
             restype=None
         )
 
         self.GenerateAndSerializeRotationKey = LattigoFunction(
             self.lib.GenerateAndSerializeRotationKey,
-            argtypes=[ctypes.c_int],
+            argtypes=[ctypes.c_ulong],
             restype=ArrayResultByte
         )
